Split up piece class to enums
