+++ PracticeLogger/ViewModel/SearchViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  SearchViewModel.swift
+//  PracticeLogger
+//
+//  Created by Michael Brandt on 7/30/24.
+//
+
+import Foundation
