+++ PracticeLogger/Views/Components/Custom Tab Bar/CustomKeyboard.swift
@@ -1,0 +1,18 @@
+//
+//  CustomKeyboard.swift
+//  PracticeLogger
+//
+//  Created by Michael Brandt on 7/30/24.
+//
+
+import SwiftUI
+
+struct CustomKeyboard: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    CustomKeyboard()
+}
