+++ PracticeLogger/Views/RepetoireRow.swift
@@ -1,0 +1,18 @@
+//
+//  RepetoireRow.swift
+//  PracticeLogger
+//
+//  Created by Michael Brandt on 7/28/24.
+//
+
+import SwiftUI
+
+struct RepetoireRow: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    RepetoireRow()
+}
