+++ PracticeLogger/model/Piece.swift
@@ -434,101 +434,3 @@ func mapResponseToFullPiece(response: [SupabasePieceResponse]) -> [Piece] {
 
     return pieces
 }
-
-enum CatalogueType: String, Decodable, CaseIterable, Encodable {
-    case B, BWV, CPEB, D, DD, EG, FMW, H, K, L, Op, S, T, TH, VB, WAB, WD, WoO, Wq
-
-    static var allCases: [CatalogueType] {
-        return [
-            .B, .BWV, .CPEB, .D, .DD, .EG, .FMW, .H, .K, .L, .Op, .S, .T, .TH, .VB, .WAB, .WD, .WoO, .Wq
-        ]
-    }
-}
-
-enum Format: String, Decodable, Encodable, CaseIterable {
-    case bagatelle = "Bagatelle"
-    case ballade = "Ballade"
-    case canon = "Canon"
-    case caprice = "Caprice"
-    case chorale = "Chorale"
-    case concerto = "Concerto"
-    case dance = "Dance"
-    case etude = "Etude"
-    case fantasy = "Fantasy"
-    case fugue = "Fugue"
-    case gavotte = "Gavotte"
-    case gigue = "Gigue"
-    case impromptu = "Impromptu"
-    case intermezzo = "Intermezzo"
-    case lied = "Lied"
-    case march = "March"
-    case mazurka = "Mazurka"
-    case mass = "Mass"
-    case minuet = "Minuet"
-    case nocturne = "Nocturne"
-    case overture = "Overture"
-    case opera = "Opera"
-    case oratorio = "Oratorio"
-    case pastiche = "Pastiche"
-    case prelude = "Prelude"
-    case polonaise = "Polonaise"
-    case rhapsody = "Rhapsody"
-    case requiem = "Requiem"
-    case rondo = "Rondo"
-    case sarabande = "Sarabande"
-    case scherzo = "Scherzo"
-    case serenade = "Serenade"
-    case sonata = "Sonata"
-    case stringQuartet = "String Quartet"
-    case suite = "Suite"
-    case symphony = "Symphony"
-    case tarantella = "Tarantella"
-    case toccata = "Toccata"
-    case variations = "Variations"
-    case waltz = "Waltz"
-}
-
-enum KeySignatureType: String, Decodable, Encodable, CaseIterable {
-    case c = "C"
-    case cSharp = "C♯"
-    case cFlat = "C♭"
-    case d = "D"
-    case dSharp = "D♯"
-    case dFlat = "D♭"
-    case e = "E"
-    case eSharp = "E♯"
-    case eFlat = "E♭"
-    case f = "F"
-    case fSharp = "F♯"
-    case fFlat = "F♭"
-    case g = "G"
-    case gSharp = "G♯"
-    case gFlat = "G♭"
-    case a = "A"
-    case aSharp = "A♯"
-    case aFlat = "A♭"
-    case b = "B"
-    case bSharp = "B♯"
-    case bFlat = "B♭"
-
-    static var allCases: [KeySignatureType] {
-        return [
-            .c, .cSharp, .cFlat,
-            .d, .dSharp, .dFlat,
-            .e, .eSharp, .eFlat,
-            .f, .fSharp, .fFlat,
-            .g, .gSharp, .gFlat,
-            .a, .aSharp, .aFlat,
-            .b, .bSharp, .bFlat
-        ]
-    }
-}
-
-enum KeySignatureTonality: String, Decodable, Encodable, CaseIterable {
-    case major = "Major"
-    case minor = "Minor"
-
-    static var allCases: [KeySignatureTonality] {
-        return [.major, .minor]
-    }
-}
